Rename message to be uppercase
